Make it work when the total size fo the array is > 32GB

+++ handle_field.cpp
@@ -4,20 +4,20 @@
 /** Functions for working with an interpolated field*/
 
 /* Multiply all elements of size field by a top hat function H(x > thresh)*/
-void multiply_by_tophat(double * field, int size, double thresh)
+void multiply_by_tophat(double * field, size_t size, double thresh)
 {
     #pragma omp parallel for
-    for(int i=0; i< size; i++)
+    for(size_t i=0; i< size; i++)
     {
         field[i] *= (field[i] > thresh);
     }
 }
 
 /* Make all positive array values 1*/
-void discretize(double * field, int size)
+void discretize(double * field, size_t size)
 {
     #pragma omp parallel for
-    for(int i=0; i< size; i++)
+    for(size_t i=0; i< size; i++)
     {
         if(field[i] > 0)
             field[i] = 1;
@@ -28,22 +28,22 @@ void discretize(double * field, int size)
 /* Normalise by the mean of the field, so that the field has mean 0,
  * sd. 1.
  * ie, calculate delta = rho / rho_bar - 1 */
-void calc_delta(double * field, int size, int realsize)
+void calc_delta(double * field, size_t size, long realsize)
 {
     const double mean = find_total(field, size) / realsize;
     #pragma omp parallel for
-    for(int i=0; i< size; i++)
+    for(size_t i=0; i< size; i++)
     {
         field[i] = field[i]/ mean -1;
     }
 }
 
 /* Find the total value of all elements of a field*/
-double find_total(double * field, int size)
+double find_total(double * field, size_t size)
 {
     double total=0;
     #pragma omp parallel for reduction(+:total)
-    for(int i=0; i< size; i++)
+    for(size_t i=0; i< size; i++)
     {
         total += field[i];
     }
@@ -57,7 +57,7 @@ double find_total(double * field, int size)
  * xmax: largest histogram bin
  * nxbins: number of bins to cover the above range. Bins are log spaced.
  *  */
-std::map<double, int> histogram(const double * field, const int size, const double xmin, const double xmax, const int nxbins)
+std::map<double, int> histogram(const double * field, const size_t size, const double xmin, const double xmax, const int nxbins)
 {
     std::map<double, int> hist;
     
@@ -67,7 +67,7 @@ std::map<double, int> histogram(const double * field, const int size, const doub
     
     //Count values: note this means that the last entry will contain the 
     //no. of elements past the end
-    for (int i = 0; i < size; ++i)
+    for (size_t i = 0; i < size; ++i)
     {
         //Lower bound = first key equal to or greater than.
         std::map<double,int>::iterator it = hist.lower_bound(*(field+i));
@@ -83,7 +83,7 @@ std::map<double, int> histogram(const double * field, const int size, const doub
 
 /** Convert a histogram to a pdf, adjusting the bins to be the bin centroids.
  */
-std::map<double, double> pdf(std::map<double, int> hist, const int size)
+std::map<double, double> pdf(std::map<double, int> hist, const size_t size)
 {
     //Convert to a pdf
     //First, calculate bins
