+++ moments.h
@@ -15,12 +15,12 @@
 //proton mass in g
 #define protonmass 1.66053886e-24
 
-void multiply_by_tophat(double * field, int size, double thresh);
-void discretize(double * field, int size);
-double find_total(double * field, int size);
-std::map<double, double> pdf(std::map<double, int> hist, const int size);
-std::map<double, int> histogram(const double * field, const int size, const double xmin, const double xmax, const int nxbins);
-void calc_delta(double * field, int size, int realsize);
+void multiply_by_tophat(double * field, size_t size, double thresh);
+void discretize(double * field, size_t size);
+double find_total(double * field, size_t size);
+std::map<double, double> pdf(std::map<double, int> hist, const size_t size);
+std::map<double, int> histogram(const double * field, const size_t size, const double xmin, const double xmax, const int nxbins);
+void calc_delta(double * field, size_t size, long realsize);
 
 #ifndef N_TYPE
         #define N_TYPE 6
@@ -42,7 +42,7 @@ class H5Snap
         int load_hdf5_dm_snapshot(const char *ffname, int fileno, int parttype, float **Pos, float **Mass);
 };
 
-int SPH_interpolate(double * field, double * comp, const int nx, float *pos, float *radii, float *value, float *weights, const double box, const int nval, const int periodic);
+int SPH_interpolate(double * field, double * comp, const int nx, float *pos, float *radii, float *value, float *weights, const double box, const long nval, const int periodic);
 int CiC_interpolate(double boxsize, int dims, double *out, int64_t segment_particles, float *positions,float *mass, int extra);
 
 #ifdef __cplusplus
